refactor: 패키치 위치 변경
- dispatcherServlet 위치 변경
- tobe 제거
- 미사용 테스트 샘플 제거
