+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -5,6 +5,7 @@ import jakarta.servlet.ServletRegistration.Dynamic;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import com.interface21.web.WebApplicationInitializer;
+import com.interface21.webmvc.servlet.mvc.DispatcherServlet;
 
 /**
  * Base class for {@link WebApplicationInitializer}
