--- app/src/test/java/samples/TestController.java
@@ -1,11 +1,0 @@
-package samples;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class TestController {
-
-    public String execute(HttpServletRequest req, HttpServletResponse res) throws Exception {
-        return "/test.jsp";
-    }
-}
