+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/AnnotationHandlerAdapter.java
@@ -1,9 +1,8 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.HandlerAdapter;
 
 public class AnnotationHandlerAdapter implements HandlerAdapter {
