+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/HandlerAdapters.java
@@ -3,7 +3,6 @@ package com.interface21.webmvc.servlet.mvc;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerAdapter;
 
 public class HandlerAdapters {
