+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/HandlerMappings.java
@@ -5,7 +5,6 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import jakarta.servlet.http.HttpServletRequest;
-import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
 
 public class HandlerMappings {
