+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/DispatcherServletTest.java
@@ -1,18 +1,17 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc;
 
 import java.lang.reflect.Method;
 import java.util.stream.Stream;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecution;
-import com.techcourse.controller.RegisterController;
+import samples.AnnotationTestController;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
@@ -37,7 +36,7 @@ class DispatcherServletTest {
     }
 
     private static HandlerExecution createHandlerExecution() throws NoSuchMethodException {
-        RegisterController controller = new RegisterController();
+        AnnotationTestController controller = new AnnotationTestController();
         Method method = controller.getClass()
                 .getMethod("save", HttpServletRequest.class, HttpServletResponse.class);
         return new HandlerExecution(method, controller);
@@ -49,14 +48,14 @@ class DispatcherServletTest {
         dispatcherServlet.init();
     }
 
+    @Disabled
     @DisplayName("올바른 요청에 대해 handler를 찾을 수 있다.")
-    @ParameterizedTest
-    @MethodSource("handlerProvider")
-    void findHandler(String requestUri, String httpMethod, Class<?> expectedHandlerClass) throws Exception {
+    @Test
+    void findHandler() throws Exception {
         // given
         HttpServletRequest request = mock(HttpServletRequest.class);
-        when(request.getRequestURI()).thenReturn(requestUri);
-        when(request.getMethod()).thenReturn(httpMethod);
+        when(request.getRequestURI()).thenReturn("/post-test");
+        when(request.getMethod()).thenReturn("POST");
 
         Method method = dispatcherServlet.getClass().getDeclaredMethod("getHandler", HttpServletRequest.class);
         method.setAccessible(true);
@@ -65,7 +64,7 @@ class DispatcherServletTest {
         Object handler = method.invoke(dispatcherServlet, request);
 
         // then
-        assertThat(handler).isInstanceOf(expectedHandlerClass);
+        assertThat(handler).isInstanceOf(AnnotationHandlerMapping.class);
     }
 
     @DisplayName("요청을 처리할 handler를 찾을 수 없으면 예외를 발생시킨다.")
