+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/HandlerKeyTest.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
